requirements.txt
Django>=5.0
psycopg2-binary>=2.9.3
gunicorn>=20.1.0
python-dotenv>=0.19.2
whitenoise>=6.0.0
dj-database-url>=0.5.0
djangorestframework>=3.14.0
